+++ tests/common_test.py
@@ -1,18 +1,20 @@
+# -*- coding: utf-8 -*-
 import mock
 
 from fido.common import listify_headers
 
 
-def test_header_value_if_list_is_not_modified():
+def test_list_value_not_listified():
     header = {'foo': ['bla']}
+    new_header = {'foo': [b'bla']}
     with mock.patch('fido.common.Headers') as mock_header:
         listify_headers(header)
-    mock_header.assert_called_once_with(header)
+    mock_header.assert_called_once_with(new_header)
 
 
-def test_header_value_if_not_list_gets_changed_to_list():
+def test_header_value_listified():
     header = {'foo': 'bla'}
-    new_header = {'foo': ['bla']}
+    new_header = {'foo': [b'bla']}
     with mock.patch('fido.common.Headers') as mock_header:
         listify_headers(header)
     mock_header.assert_called_once_with(new_header)