+++ tests/fido_test.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import json
 import logging
 import mock
 import threading
@@ -15,16 +14,18 @@ import twisted.web.client
 from twisted.web.client import Agent
 from twisted.web.client import FileBodyProducer
 from twisted.web.client import ProxyAgent
+from yelp_bytes import to_bytes
 
 import fido
 from fido.fido import _build_body_producer
 from fido.fido import _set_deferred_timeout
-from fido.fido import _url_to_utf8
+from fido.fido import DEFAULT_USER_AGENT
 
 SERVER_OVERHEAD_TIME = 2.0
 TIMEOUT_TEST = 1.0
 
 ERROR_MESSAGE = 'I failed :('
+ECHO_URL = '/echo'
 
 
 @pytest.yield_fixture(scope="module")
@@ -35,31 +36,47 @@ def server_url():
     logging.basicConfig()
     logging.getLogger('twisted').setLevel(logging.CRITICAL)
 
-    class MyHandler(BaseHTTPServer.BaseHTTPRequestHandler):
+    class TestHandler(BaseHTTPServer.BaseHTTPRequestHandler):
         def echo(self):
-            if '/slow' in self.path:
+            if 'slow' in self.path:
                 time.sleep(SERVER_OVERHEAD_TIME)
+
             self.send_response(200)
-            self.send_header("Content-type", "application/json")
+
+            for k, v in self.headers.items():
+                self.send_header(k, v)
             self.end_headers()
+
             content_length = int(self.headers.get('Content-Length', 0))
-            self.wfile.write(json.dumps({
-                'headers': dict(self.headers),
-                'method': self.command,
-                'body': self.rfile.read(content_length)
-            }))
+            if content_length > 0:
+                self.wfile.write(self.rfile.read(content_length))
+
+        def content_length(self):
+            """Send back the content-length number as the response."""
+            self.send_response(200)
+
+            response = to_bytes(self.headers.get('Content-Length'))
+            content_length = len(response)
+            self.send_header('Content-Length', content_length)
+            self.end_headers()
+
+            self.wfile.write(response)
 
         def do_GET(self):
-            self.echo()
+            if ECHO_URL in self.path:
+                self.echo()
 
         def do_POST(self):
-            self.echo()
+            if 'content_length' in self.path:
+                self.content_length()
+            elif ECHO_URL in self.path:
+                self.echo()
 
     class MultiThreadedHTTPServer(SocketServer.ThreadingMixIn,
                                   BaseHTTPServer.HTTPServer):
         request_queue_size = 1000
 
-    httpd = MultiThreadedHTTPServer(('localhost', 0), MyHandler)
+    httpd = MultiThreadedHTTPServer(('localhost', 0), TestHandler)
     httpd_thread = threading.Thread(target=httpd.serve_forever)
     httpd_thread.start()
     yield 'http://%s:%d/' % (httpd.server_address[0], httpd.server_address[1])
@@ -67,15 +84,12 @@ def server_url():
     httpd_thread.join()
 
 
-def test_unicode_url():
-    assert _url_to_utf8(u'繁') == '\xe7\xb9\x81'
-
-
 def test_fetch_basic(server_url):
-    response = fido.fetch(server_url).wait()
-    assert 'Content-Type' in response.headers
+    response = fido.fetch(server_url + ECHO_URL).wait()
+    assert response.headers.get(b'User-Agent') == \
+        [to_bytes(DEFAULT_USER_AGENT)]
+    assert response.reason == b'OK'
     assert response.code == 200
-    assert response.json()['method'] == 'GET'
 
 
 def test_set_deferred_timeout_none():
@@ -106,7 +120,7 @@ def test_eventual_result_timeout(server_url):
     """
 
     # fetch without setting timeouts -> we could potentially wait forever
-    eventual_result = fido.fetch(server_url + 'slow')
+    eventual_result = fido.fetch(server_url + ECHO_URL + '/slow')
 
     # make sure no timeout error is thrown here but only on result retrieval
     assert eventual_result.original_failure() is None
@@ -122,7 +136,10 @@ def test_agent_timeout(server_url):
     Testing that we don't wait forever on the server sending back a response
     """
 
-    eventual_result = fido.fetch(server_url + 'slow', timeout=TIMEOUT_TEST)
+    eventual_result = fido.fetch(
+        server_url + ECHO_URL + '/slow',
+        timeout=TIMEOUT_TEST
+    )
 
     # wait for fido to estinguish the timeout and abort before test-assertions
     time.sleep(2 * TIMEOUT_TEST)
@@ -142,7 +159,7 @@ def test_agent_timeout(server_url):
     )
 
 
-def test_agent_connect_timeout(server_url):
+def test_agent_connect_timeout():
     """
     Testing that we don't wait more than connect_timeout to establish a http
     connection
@@ -173,27 +190,20 @@ def test_agent_connect_timeout(server_url):
 
 def test_fetch_stress(server_url):
     eventual_results = [
-        fido.fetch(server_url, timeout=8) for _ in range(1000)
+        fido.fetch(server_url + ECHO_URL, timeout=10) for _ in range(1000)
     ]
     for eventual_result in eventual_results:
         eventual_result.wait(timeout=10)
 
 
-def test_fetch_method(server_url):
-    expected_method = 'POST'
-    eventual_result = fido.fetch(server_url, method=expected_method)
-    actual_method = eventual_result.wait().json()['method']
-    assert expected_method == actual_method
-
-
 def test_fetch_headers(server_url):
     headers = {'foo': ['bar']}
-    eventual_result = fido.fetch(server_url, headers=headers)
-    actual_headers = eventual_result.wait().json()['headers']
-    assert actual_headers.get('foo') == 'bar'
+    eventual_result = fido.fetch(server_url + ECHO_URL, headers=headers)
+    actual_headers = eventual_result.wait().headers
+    assert actual_headers.get(b'Foo') == [b'bar']
 
 
-def test_headers_keep_content_length_if_no_body(server_url):
+def test_headers_keep_content_length_if_no_body():
     headers = {'Content-Length': '0'}
     bodyProducer, headers = _build_body_producer(None, headers)
     assert bodyProducer is None
@@ -203,54 +213,78 @@ def test_headers_keep_content_length_if_no_body(server_url):
 @pytest.mark.parametrize(
     'header', ('content-length', 'Content-Length')
 )
-def test_headers_remove_content_length_if_body(server_url, header):
+def test_headers_remove_content_length_if_body(header):
     headers = {header: '22'}
-    body = '{"some_json_data": 30}'
+    body = b'{"some_json_data": 30}'
 
     bodyProducer, headers = _build_body_producer(body, headers)
     assert isinstance(bodyProducer, FileBodyProducer)
     assert headers == {}
 
 
+def test_json_body(server_url):
+    body = b'{"some_json_data": 30}'
+    eventual_result = fido.fetch(
+        server_url + ECHO_URL,
+        method='POST',
+        body=body
+    )
+    assert eventual_result.wait().json()['some_json_data'] == 30
+
+
 def test_content_length_readded_by_twisted(server_url):
-    headers = {'Content-Length': '22'}
-    body = '{"some_json_data": 30}'
-    eventual_result = fido.fetch(server_url, headers=headers, body=body)
-    actual_headers = eventual_result.wait().json()['headers']
-    assert actual_headers.get('content-length') == '22'
+    headers = {'Content-Length': '250'}
+    body = b'{"some_json_data": 30}'
+    eventual_result = fido.fetch(
+        server_url + '/content_length',
+        method='POST',
+        headers=headers,
+        body=body
+    )
+    content_length = int(eventual_result.wait().body)
+    assert content_length == 22
 
 
-def test_headers_not_modified_in_place(server_url):
+def test_headers_not_modified_in_place():
     headers = {'foo': 'bar', 'Content-Length': '22'}
-    body = '{"some_json_data": 30}'
+    body = b'{"some_json_data": 30}'
     _, _ = _build_body_producer(body, headers)
 
     assert headers == {'foo': 'bar', 'Content-Length': '22'}
 
 
 def test_fetch_content_type(server_url):
-    expected_content_type = 'text/html'
+    expected_content_type = b'text/html'
     eventual_result = fido.fetch(
-        server_url,
+        server_url + ECHO_URL,
         headers={'Content-Type': expected_content_type}
     )
-    actual_content_type = eventual_result.wait().json()['headers'].\
-        get('content-type')
-    assert expected_content_type == actual_content_type
+    actual_content_type = eventual_result.wait().headers.\
+        get(b'Content-Type')
+    assert [expected_content_type] == actual_content_type
 
 
-def test_fetch_user_agent(server_url):
-    expected_user_agent = 'skynet'
-    eventual_result = fido.fetch(server_url, user_agent=expected_user_agent)
-    actual_user_agent = eventual_result.wait().json()['headers'].\
-        get('user-agent')
+@pytest.mark.parametrize(
+    'header_name', ('User-Agent', 'user-agent')
+)
+def test_fetch_user_agent(server_url, header_name):
+    expected_user_agent = [b'skynet']
+    headers = {header_name: expected_user_agent}
+    eventual_result = fido.fetch(
+        server_url + ECHO_URL,
+        headers=headers,
+    )
+    actual_user_agent = eventual_result.wait().headers.get(b'User-Agent')
     assert expected_user_agent == actual_user_agent
 
 
 def test_fetch_body(server_url):
-    expected_body = 'corpus'
-    eventual_result = fido.fetch(server_url, body=expected_body)
-    actual_body = eventual_result.wait().json()['body']
+    expected_body = b'corpus'
+    eventual_result = fido.fetch(
+        server_url + ECHO_URL,
+        body=expected_body
+    )
+    actual_body = eventual_result.wait().body
     assert expected_body == actual_body
 
 
@@ -289,7 +323,7 @@ def test_deferred_errback_chain():
         with pytest.raises(twisted.web.client.ResponseFailed) as e:
             eventual_result.wait(timeout=TIMEOUT_TEST)
 
-        assert e.value.message == ERROR_MESSAGE
+        assert e.value.args == (ERROR_MESSAGE,)
 
 
 def test_fetch_inner_exception_thrown_in_reactor_thread():
@@ -314,7 +348,7 @@ def test_fetch_inner_exception_thrown_in_reactor_thread():
             with pytest.raises(ValueError) as e:
                 eventual_result.wait(timeout=TIMEOUT_TEST)
 
-            assert e.value.message == ERROR_MESSAGE
+            assert e.value.args == (ERROR_MESSAGE,)
 
 
 def test_fetch_normally_throws_exception():